feat: :art: Formato de return de rotas de listagem alterado; criado decorator de sessão do usuario; criado guard p/ autenticar token jwt; propriedade updated at adicionada nos models

+++ src/controllers/user.controller.ts
@@ -1,10 +1,13 @@
-import { Controller, Get, Delete, Param, Body, Put, Query } from '@nestjs/common';
+import { Controller, Get, Delete, Param, Body, Put, Query, UseGuards, ForbiddenException } from '@nestjs/common';
 import { ApiTags } from '@nestjs/swagger';
 import { UserService } from '../services/user.service';
 import { UpdateUserDto } from '../dtos/user/update-user.dto';
+import { JwtAuthGuard } from 'src/guards/auth.guard';
+import { SessionUser } from 'src/decorators/sessionUser.decorator';
 
 @ApiTags('users')
 @Controller('user')
+@UseGuards(JwtAuthGuard)
 export class UserController {
   constructor(private readonly usersService: UserService) {}
 
@@ -13,7 +16,7 @@ export class UserController {
     @Query('page') page?: string,
     @Query('limit') limit?: string,
   ) {
-    return this.usersService.findAll(Number(page) || 1, Number(limit) || 20);
+    return this.usersService.findAll(Number(page) || 1, Number(limit) || 100);
   }
 
   @Get(':id')
@@ -22,12 +25,22 @@ export class UserController {
   }
 
   @Put(':id')
-  update(@Param('id') id: string, @Body() data: UpdateUserDto) {
+  update(@Param('id') id: string, @Body() data: UpdateUserDto, @SessionUser() user: any) {
+    if (
+      user.role !== 'admin'
+      && user.sub !== id
+    ) throw new ForbiddenException('Access denied');
+
     return this.usersService.update(id, data);
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  remove(@Param('id') id: string, @SessionUser() user: any) {
+    if (
+      user.role !== 'admin'
+      && user.sub !== id
+    ) throw new ForbiddenException('Access denied');
+  
     return this.usersService.delete(id);
   }
 }
